refactor(client): reduce dashboard route & card imports via shared barrels
The dashboard route and its tile cards tripped import/max-dependencies (10):
each card re-imported the same card chrome from five separate modules, the
route file hand-rolled the layout-tab menu inline, and the grid imported all
eleven tile components individually.

Introduce three route-private barrels so each consumer imports one surface:
- -cardKit re-exports the primitives every card shares (DashboardCard, settings
  flyout, tabs, popover, radial progress, cn, queryKeys)
- -tileComponents owns the tile registry the grid renders
- index.ts composes the route's pieces plus shared dialogs/chrome
Also extract the inline LayoutTab into -LayoutTab. All ten tile cards migrate
to -cardKit for consistency.

Behavior-only-equivalent: every change is an import-source swap or a mechanical
component extraction. All six files in #306 now pass import/max-dependencies;
typecheck and the client test suite stay green.

Resolves #306.
