+++ packages/client/src/routes/dashboard.-components/-DashboardChangelog.tsx
@@ -9,11 +9,12 @@ import type { ReactNode } from "react";
 // fallow-ignore-next-line unresolved-import
 import changelogMarkdown from "@root/CHANGELOG.md?raw";
 
+import {
+  CardSettingsFlyout,
+  DashboardCard,
+  isAutoHeight,
+} from "./-cardKit";
 import { parseChangelog } from "./-changelog";
-import { CardSettingsFlyout } from "./-DashboardCardSettings";
-import { isAutoHeight } from "./-dashboardTileMeta";
-
-import { DashboardCard } from "@/components/boxes/DashboardCard";
 
 const REPO_URL = "https://github.com/emilyeserven/course-tracker";
