+++ packages/client/src/routes/dashboard.-components/-DashboardUnderutilizedProviders.tsx
@@ -6,14 +6,14 @@ import { useQuery } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
 import { ExternalLink } from "lucide-react";
 
-import { CardSettingsFlyout } from "./-DashboardCardSettings";
-import { isAutoHeight } from "./-dashboardTileMeta";
-
 import {
+  Button,
+  CardSettingsFlyout,
   DashboardCard,
   DashboardSectionStatus,
-} from "@/components/boxes/DashboardCard";
-import { Button } from "@/components/ui/button";
+  isAutoHeight,
+} from "./-cardKit";
+
 import { DataTable } from "@/components/ui/data-table";
 import { fetchProviders, formatCurrency, parseCost } from "@/utils";
