+++ packages/client/src/routes/dashboard.-components/-cardKit.ts
@@ -1,0 +1,16 @@
+// Shared primitives for the dashboard tile cards, so each card imports the
+// common surface (card chrome, settings flyout, tabs, popover, progress) from
+// one module instead of five. Cards import this barrel; the sources below
+// import their own primitives directly, so this re-export introduces no cycle.
+export {
+  DashboardCard,
+  DashboardSectionStatus,
+} from "@/components/boxes/DashboardCard";
+export { CardSettingsFlyout, SettingToggle } from "./-DashboardCardSettings";
+export { isAutoHeight } from "./-dashboardTileMeta";
+export { Button } from "@/components/ui/button";
+export { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
+export { Popover, PopoverContent, PopoverTrigger } from "@/components/popover";
+export { RadialProgress } from "@/components/ui/RadialProgress";
+export { cn } from "@/lib/utils";
+export { queryKeys } from "@/utils/queryKeys";
